LayoutTests/platform/ios/imported/w3c/web-platform-tests/dom/events/scrolling/scrollend-event-fired-for-scrollIntoView_include=root-inline-end-block-end-behavior-auto-expected.txt


FAIL Tests scrollend event for scrollIntoView with behavior 'auto' on root. assert_equals: HTML.scrollIntoView scrollLeft expected 10000 but got 2200
